# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ControlActionsModel": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ControlActionModel"
            },
            "title": "Actions",
            "type": "array"
          }
        },
        "required": [
          "actions"
        ],
        "title": "ControlActionsModel",
        "type": "object"
      },
...
      "GuiDelayRequestModel": {
        "properties": {
          "delay_ms": {
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Delay Ms",
            "type": "integer"
          }
        },
        "required": [
          "delay_ms"
        ],
        "title": "GuiDelayRequestModel",
        "type": "object"
      },
      "GuiDelayResponseModel": {
        "properties": {
          "delay_ms": {
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Delay Ms",
            "type": "integer"
          }
        },
        "required": [
          "delay_ms"
        ],
        "title": "GuiDelayResponseModel",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NativeGuiResponseModel": {
        "properties": {
          "status": {
            "const": "shown",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "NativeGuiResponseModel",
        "type": "object"
      },
...
      "PredictionResultModel": {
        "properties": {
          "horizon_seconds": {
            "title": "Horizon Seconds",
            "type": "number"
          },
          "horizon_steps": {
            "title": "Horizon Steps",
            "type": "integer"
          },
          "predicted_flows": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Predicted Flows",
            "type": "object"
          }
        },
        "required": [
          "horizon_steps",
          "horizon_seconds",
          "predicted_flows"
        ],
        "title": "PredictionResultModel",
        "type": "object"
      },
...
      "ResultDetailModel": {
        "description": "Validated result payload that never exposes a filesystem path.",
        "properties": {
          "algorithm": {
            "default": "",
            "title": "Algorithm",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scene_id": {
            "title": "Scene Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "run_id",
          "status",
          "reason",
          "scene_id",
          "summary"
        ],
        "title": "ResultDetailModel",
        "type": "object"
      },
...
      "RunResultModel": {
        "properties": {
          "algorithm": {
            "default": "",
            "title": "Algorithm",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "run_dir": {
            "title": "Run Dir",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "run_id",
          "status",
          "reason",
          "run_dir"
        ],
        "title": "RunResultModel",
        "type": "object"
      },
...
      "StateRequestModel": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/JointStateModel"
          }
        },
        "required": [
          "state"
        ],
        "title": "StateRequestModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "雄安车路云协同管控平台",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cloud/predict": {
      "post": {
        "operationId": "cloud_predict_api_cloud_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cloud Predict"
      }
    },
    "/api/edge/control": {
      "post": {
        "operationId": "edge_control_api_edge_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlActionsModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edge Control"
      }
    },
...
    "/api/results/{run_id}": {
      "get": {
        "operationId": "get_result_api_results__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultDetailModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "validated result unavailable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result"
      }
    },
    "/api/runs": {
      "post": {
        "operationId": "submit_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Run"
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/api/runs/{run_id}/frame": {
      "get": {
        "operationId": "get_frame_api_runs__run_id__frame_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sequence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sequence"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Frame-Sequence": {
                "schema": {
                  "type": "integer"
                }
              },
              "X-Run-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Simulation-Time": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "404": {
            "description": "run or frame unavailable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Frame"
      }
    },
    "/api/runs/{run_id}/gui-delay": {
      "put": {
        "operationId": "set_gui_delay_api_runs__run_id__gui_delay_put",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuiDelayRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuiDelayResponseModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown run_id"
          },
          "409": {
            "description": "GUI delay unavailable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Gui Delay"
      }
    },
...
    "/api/runs/{run_id}/native-gui": {
      "post": {
        "operationId": "show_native_gui_api_runs__run_id__native_gui_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NativeGuiResponseModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown run_id"
          },
          "409": {
            "description": "native SUMO-GUI unavailable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show Native Gui"
      }
    },
    "/api/runs/{run_id}/safety": {
      "get": {
        "operationId": "get_safety_api_runs__run_id__safety_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "validated safety unavailable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Safety"
      }
    },
    "/api/runs/{run_id}/stop": {
      "post": {
        "operationId": "stop_run_api_runs__run_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Run"
      }
    },
...
    "/api/simulation/start": {
      "post": {
        "deprecated": true,
        "operationId": "legacy_simulation_start_api_simulation_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyRunRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Legacy Simulation Start"
      }
    },
    "/api/simulation/stop": {
      "post": {
        "deprecated": true,
        "operationId": "legacy_simulation_stop_api_simulation_stop_post",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Legacy Simulation Stop"
      }
    },
...
    "/run": {
      "post": {
        "deprecated": true,
        "operationId": "legacy_run_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyRunRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Legacy Run"
      }
    },
...
    "/status": {
      "get": {
        "deprecated": true,
        "operationId": "legacy_status_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Legacy Status"
      }
    }
  },
...
}